backend/requirements.txt
flask==2.1.0
flask-cors==3.0.10
instaloader==4.9.5
openai==1.3.0
requests==2.26.0
python-dotenv==0.19.2